added check to ensure something is typed in

+++ admin/__init__.py
@@ -25,59 +25,67 @@ def admin_view_results():
 def add_question():
     section = 1
     if section == 1:
-        question = [Questions(question = request.form['question'],
+        if 'question' in request.form and 'answer' in request.form:
+            question = [Questions(question = request.form['question'],
                                    answer = request.form['answer'],
                                    is_Correct = True,
                                    section = 1)
                                    ]
-        database_session.add(question)
-        database_session.commit()
+            database_session.add(question)
+            database_session.commit()
     elif section == 2:
-        question = [Questions(question = request.form['question'],
+        if 'question' in request.form and 'answer' in request.form:
+            question = [Questions(question = request.form['question'],
                                     answer = "this is a section 2 question",
                                     is_Correct = True,
                                     section = 2)]
-        database_session.add(question)
-        database_session.commit()
+            database_session.add(question)
+            database_session.commit()
     elif section == 3:
-        question = [Questions(question = request.form['question'],
+        if 'question' in request.form and 'answer' in request.form:
+            question = [Questions(question = request.form['question'],
                                     answer = "this is a section 3 question",
                                     is_Correct = True,
                                     section = 3)]
-        database_session.add(question)
-        database_session.commit()
+            database_session.add(question)
+            database_session.commit()
 
 
 def add_answer():
-    question = [Questions(question = request.form['question'],
+    if 'question' in request.form and 'answer' in request.form:
+        question = [Questions(question = request.form['question'],
                                 answer = request.form['answer'],
                                 is_Correct = False,
                                 section = 1)
                                 ]
-    database_session.add(question)
-    database_session.commit()
+        database_session.add(question)
+        database_session.commit()
 
 def delete_question():
-    del_query = database_session.query(Questions).filter(Questions.question==request.form['question'])
-    del_query.delete()
-    database_session.commit()
+    if 'question' in request.form:
+        del_query = database_session.query(Questions).filter(Questions.question==request.form['question'])
+        del_query.delete()
+        database_session.commit()
 
 def delete_answer():
-    del_query = database_session.query(Questions).filter(and_(Questions.question==request.form['question'] , Questions.answer==request.form['answer']))
-    del_query.delete()
-    database_session.commit()
+    if 'question' in request.form and 'answer' in request.form:
+        del_query = database_session.query(Questions).filter(and_(Questions.question==request.form['question'] , Questions.answer==request.form['answer']))
+        del_query.delete()
+        database_session.commit()
 
 
 def edit_question():
-    rows_to_update = database_session.query(Questions).filter(Questions.question == request.form['question'])
-    for row in rows_to_update:
-        row.question = request.form['new_question']
-    database_session.commit()
+    if 'question' in request.form and 'new_question' in request.form:
+        rows_to_update = database_session.query(Questions).filter(Questions.question == request.form['question'])
+        for row in rows_to_update:
+            row.question = request.form['new_question']
+        database_session.commit()
 
 def edit_answer():
-    rows_to_update = database_session.query(Questions).filter(and_(Questions.question == request.form['question'] , Questions.answer == request.form['answer']))
-        rows_to_update.answer=request.form['new_answer']
-    database_session.commit()
+    if 'question' in request.form and 'answer' in request.form and 'new_answer' in request.form:
+        rows_to_update = database_session.query(Questions).filter(and_(Questions.question == request.form['question'] , Questions.answer == request.form['answer']))
+            rows_to_update.answer=request.form['new_answer']
+        database_session.commit()
 
 def clear_student_answers():
     del_query = database_session.query(StudentAnswers)
